requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.3.2
pymongo==4.6.3
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
